requirements.txt
# Core

# Web scraping
selenium==4.18.1
newspaper3k==0.2.8

# NLP & Embeddings
sentence-transformers==2.5.1
scikit-learn==1.3.2
numpy==1.24.4

# Vector search
faiss-cpu==1.7.4

# LLM
torch==2.1.2
transformers==4.39.3
accelerate==0.27.2

# Utilities
tqdm==4.66.2
